+++ Tests/Unit/Search/IndexMetadataTest.php
@@ -15,7 +15,10 @@ use Massive\Bundle\SearchBundle\Search\Metadata\IndexMetadata;
 
 class IndexMetadataTest extends \PHPUnit_Framework_TestCase
 {
-    protected $indexMetadata;
+    /**
+     * @var IndexMetadata
+     */
+    private $metadata;
 
     public function setUp()
     {
@@ -26,8 +29,5 @@ class IndexMetadataTest extends \PHPUnit_Framework_TestCase
     {
         $this->metadata->setImageUrlField('field');
         $this->assertEquals('field', $this->metadata->getImageUrlField());
-
-        $this->metadata->setCategoryName('cat_1');
-        $this->assertEquals('cat_1', $this->metadata->getCategoryName());
     }
 }