+++ Tests/Unit/Search/Metadata/Driver/XmlDriverTest.php
@@ -1,104 +1,113 @@
 <?php
 
-namespace Massive\Bundle\SearchBundle\Tests\Unit\Search\Metadata\Driver;
+/*
+ * This file is part of Sulu.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Massive\Bundle\SearchBundle\Search\Metadata\Driver;
 
 use Massive\Bundle\SearchBundle\Search\Factory;
-use Massive\Bundle\SearchBundle\Search\Metadata\Driver\XmlDriver;
+use Massive\Bundle\SearchBundle\Search\Metadata\ClassMetadata;
+use Massive\Bundle\SearchBundle\Search\Metadata\Field\Expression;
+use Massive\Bundle\SearchBundle\Search\Metadata\Field\Field;
+use Massive\Bundle\SearchBundle\Search\Metadata\Field\Property;
+use Massive\Bundle\SearchBundle\Search\Metadata\IndexMetadata;
+use Metadata\Driver\FileLocatorInterface;
+use Prophecy\Argument;
 
 class XmlDriverTest extends \PHPUnit_Framework_TestCase
 {
-    const TMP_FILE = 'mapping_test';
+    /**
+     * @var XmlDriver
+     */
+    private $xmlDriver;
+
+    /**
+     * @var Factory
+     */
+    private $factory;
+
+    /**
+     * @var FileLocatorInterface
+     */
+    private $fileLocator;
+
+    /**
+     * @var \ReflectionClass
+     */
+    private $reflectionClass;
 
     public function setUp()
     {
-        $factory = new Factory();
-        $this->locator = $this->prophesize('Metadata\Driver\FileLocatorInterface');
-        $this->driver = new XmlDriver(
-            $factory,
-            $this->locator->reveal()
-        );
+        $this->factory = $this->prophesize(Factory::class);
+        $this->fileLocator = $this->prophesize(FileLocatorInterface::class);
+
+        $this->reflectionClass = $this->prophesize(\ReflectionClass::class);
+        $this->reflectionClass->getName()->willReturn('Sulu\Bundle\ExampleBundle\Entity\Example');
+
+        $that = $this;
+
+        $this->factory->createClassMetadata(Argument::cetera())->will(function ($arguments) use ($that) {
+            $classMetadata = new ClassMetadata($that->reflectionClass->getName());
+
+            return $classMetadata;
+        });
+
+        $this->factory->createMetadataProperty(Argument::cetera())->will(function ($arguments) use ($that) {
+            return new Property($arguments[0]);
+        });
+
+        $this->factory->createMetadataField(Argument::cetera())->will(function ($arguments) use ($that) {
+            return new Field($arguments[0]);
+        });
+
+        $this->factory->createIndexMetadata()->will(function () use ($that) {
+            return new IndexMetadata();
+        });
+
+        $this->factory->createMetadataExpression(Argument::cetera())->will(function ($arguments) use ($that) {
+            return new Expression($arguments[0]);
+        });
+
+        $this->xmlDriver = new XmlDriver($this->factory->reveal(), $this->fileLocator->reveal());
     }
 
-    public function tearDown()
-    {
-        $mappingPath = $this->getTmpMappingPath();
-        if (file_exists($mappingPath)) {
-            unlink($mappingPath);
-        }
-    }
-
-    /**
-     * It should load metadata from a file.
-     */
     public function testLoadMetadataFromFile()
     {
-        $path = $this->createMapping(<<<EOT
-<massive-search-mapping xmlns="http://massiveart.com/schema/dic/massive-search-mapping">
+        $metadata = $this->xmlDriver->loadMetadataFromFile(
+            $this->reflectionClass->reveal(),
+            __DIR__ . '/../../../../Resources/DataFixtures/Mapping/Example.xml'
+        );
 
-    <mapping class="stdClass">
-        <index name="index_bicycle"/>
-        <id property="id"/>
-        <title property="title" />
-        <category name="transport" />
-        <fields>
-            <field name="title" expr="object.title" type="string" />
-        </fields>
-    </mapping>
+        $indexMetadata = $metadata->getIndexMetadata('_default');
 
-</massive-search-mapping>
-EOT
-    );
-        $reflection = new \ReflectionClass('stdClass');
-        $metadata = $this->driver->loadMetadataFromFile($reflection, $path);
-
-        $this->assertInstanceOf('Massive\Bundle\SearchBundle\Search\Metadata\ClassMetadata', $metadata);
-        $indexMetadatas = $metadata->getIndexMetadatas();
-        $this->assertCount(1, $indexMetadatas);
-        $indexMetadata = reset($indexMetadatas);
-        $this->assertEquals('index_bicycle', $indexMetadata->getIndexName());
-        $this->assertInstanceOf('Massive\Bundle\SearchBundle\Search\Metadata\Field\Property', $indexMetadata->getIdField());
         $this->assertEquals('id', $indexMetadata->getIdField()->getProperty());
         $this->assertEquals('title', $indexMetadata->getTitleField()->getProperty());
-        $this->assertEquals('transport', $indexMetadata->getCategoryName());
-        $this->assertCount(1, $indexMetadata->getFieldMapping());
+        $this->assertEquals('description', $indexMetadata->getDescriptionField()->getProperty());
+        $this->assertEquals('example', $indexMetadata->getIndexName()->getName());
+        $this->assertEquals('locale', $indexMetadata->getLocaleField()->getProperty());
+
+        $fieldMappings = $indexMetadata->getFieldMapping();
+        $this->assertEquals('string', $fieldMappings['title']['type']);
+        $this->assertEquals('title', $fieldMappings['title']['field']->getName());
+        $this->assertEquals('string', $fieldMappings['description']['type']);
+        $this->assertEquals('description', $fieldMappings['description']['field']->getName());
     }
 
-    /**
-     * It should allow the specification of reindex directives.
-     */
-    public function testReindexDirectives()
+    public function testLoadMetadataFromFileWithEvaluatingIndex()
     {
-        $path = $this->createMapping(<<<EOT
-<massive-search-mapping xmlns="http://massiveart.com/schema/dic/massive-search-mapping">
+        $metadata = $this->xmlDriver->loadMetadataFromFile(
+            $this->reflectionClass->reveal(),
+            __DIR__ . '/../../../../Resources/DataFixtures/Mapping/ExampleWithEvaluatingIndex.xml'
+        );
 
-    <mapping class="stdClass">
-        <index name="foo"/>
-        <id property="id"/>
-        <title property="foo" />
-        <reindex repository-method="findLatest" />
-        <fields/>
-    </mapping>
+        $indexMetadata = $metadata->getIndexMetadata('_default');
 
-</massive-search-mapping>
-EOT
-    );
-        $reflection = new \ReflectionClass('stdClass');
-        $metadata = $this->driver->loadMetadataFromFile($reflection, $path);
-        $this->assertEquals('findLatest', $metadata->getReindexRepositoryMethod());
-    }
-
-    private function createMapping($mapping)
-    {
-        $path = $this->getTmpMappingPath();
-        file_put_contents($path, $mapping);
-
-        return $path;
-    }
-
-    private function getTmpMappingPath()
-    {
-        $tempDir = sys_get_temp_dir();
-
-        return $tempDir . '/' . self::TMP_FILE;
+        $this->assertEquals('object.getWebspace()', $indexMetadata->getIndexName()->getExpression());
     }
 }